renamed joke categories
