+++ src/app/jokes-by-category/jokes-by-category.component.html
